refactor: use camel case for log visibility policy
