+++ packages/api/src/prividium/prividium.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
 import { PrividiumRulesService } from "./prividium-rules.service";
-import { RuleBasedLogVisibilityPolicy } from "./policies/log-visibility.policy";
+import { RuleBasedLogVisibilityPolicy } from "./policies/logVisibility.policy";
 
 @Module({
   providers: [PrividiumRulesService, RuleBasedLogVisibilityPolicy],
